+++ src/websocket/interface/custom-socket.interface.ts
